+++ src/main/java/org/torproject/descriptor/impl/RelayNetworkStatusVoteImpl.java
@@ -52,6 +52,8 @@ public class RelayNetworkStatusVoteImpl extends NetworkStatusImpl
     this.checkExactlyOnceKeywords(exactlyOnceKeywords);
     Set<String> atMostOnceKeywords = new HashSet<>(Arrays.asList((
         "consensus-methods,client-versions,server-versions,"
+        + "recommended-client-protocols,recommended-relay-protocols,"
+        + "required-client-protocols,required-relay-protocols,"
         + "flag-thresholds,params,contact,"
         + "legacy-key,dir-key-crosscert,dir-address,directory-footer")
         .split(",")));
@@ -116,6 +118,18 @@ public class RelayNetworkStatusVoteImpl extends NetworkStatusImpl
           break;
         case "server-versions":
           this.parseServerVersionsLine(line, parts);
+          break;
+        case "recommended-client-protocols":
+          this.parseRecommendedClientProtocolsLine(line, parts);
+          break;
+        case "recommended-relay-protocols":
+          this.parseRecommendedRelayProtocolsLine(line, parts);
+          break;
+        case "required-client-protocols":
+          this.parseRequiredClientProtocolsLine(line, parts);
+          break;
+        case "required-relay-protocols":
+          this.parseRequiredRelayProtocolsLine(line, parts);
           break;
         case "package":
           this.parsePackageLine(line, parts);
@@ -303,6 +317,30 @@ public class RelayNetworkStatusVoteImpl extends NetworkStatusImpl
       throws DescriptorParseException {
     this.recommendedServerVersions = this.parseClientOrServerVersions(
         line, parts);
+  }
+
+  private void parseRecommendedClientProtocolsLine(String line, String[] parts)
+      throws DescriptorParseException {
+    this.recommendedClientProtocols = ParseHelper.parseProtocolVersions(line,
+        line, parts);
+  }
+
+  private void parseRecommendedRelayProtocolsLine(String line, String[] parts)
+      throws DescriptorParseException {
+    this.recommendedRelayProtocols = ParseHelper.parseProtocolVersions(line,
+        line, parts);
+  }
+
+  private void parseRequiredClientProtocolsLine(String line, String[] parts)
+      throws DescriptorParseException {
+    this.requiredClientProtocols = ParseHelper.parseProtocolVersions(line,
+        line, parts);
+  }
+
+  private void parseRequiredRelayProtocolsLine(String line, String[] parts)
+      throws DescriptorParseException {
+    this.requiredRelayProtocols = ParseHelper.parseProtocolVersions(line, line,
+        parts);
   }
 
   private void parsePackageLine(String line, String[] parts)
@@ -708,6 +746,34 @@ public class RelayNetworkStatusVoteImpl extends NetworkStatusImpl
   public List<String> getRecommendedServerVersions() {
     return this.recommendedServerVersions == null ? null
         : Arrays.asList(this.recommendedServerVersions);
+  }
+
+  private SortedMap<String, SortedSet<Long>> recommendedClientProtocols;
+
+  @Override
+  public SortedMap<String, SortedSet<Long>> getRecommendedClientProtocols() {
+    return this.recommendedClientProtocols;
+  }
+
+  private SortedMap<String, SortedSet<Long>> recommendedRelayProtocols;
+
+  @Override
+  public SortedMap<String, SortedSet<Long>> getRecommendedRelayProtocols() {
+    return this.recommendedRelayProtocols;
+  }
+
+  private SortedMap<String, SortedSet<Long>> requiredClientProtocols;
+
+  @Override
+  public SortedMap<String, SortedSet<Long>> getRequiredClientProtocols() {
+    return this.requiredClientProtocols;
+  }
+
+  private SortedMap<String, SortedSet<Long>> requiredRelayProtocols;
+
+  @Override
+  public SortedMap<String, SortedSet<Long>> getRequiredRelayProtocols() {
+    return this.requiredRelayProtocols;
   }
 
   private List<String> packageLines;